Moved ErrorIs related code to files restricted to go1.13
