+++ assert/assertion_format.go
@@ -115,20 +115,6 @@ func Errorf(t TestingT, err error, msg string, args ...interface{}) bool {
 	return Error(t, err, append([]interface{}{msg}, args...)...)
 }
 
-// ErrorIsf asserts that a function returned an error chain
-// and that it is contains the provided error.
-// Mimicking errors.Is, providing `nil` as expected error
-// is the same as calling NoError
-//
-//   actualObj, err := SomeFunction()
-//   assert.ErrorIsf(t, err,  expectedError, "error message %s", "formatted")
-func ErrorIsf(t TestingT, expected error, actual error, msg string, args ...interface{}) bool {
-	if h, ok := t.(tHelper); ok {
-		h.Helper()
-	}
-	return ErrorIs(t, expected, actual, append([]interface{}{msg}, args...)...)
-}
-
 // Eventuallyf asserts that given condition will be met in waitFor time,
 // periodically checking target function each tick.
 //
